requirements.txt
# =============================================================================
# DEPRECATION WARNING:
#
# The file `requirements.txt` does not influence the package dependencies and
# will not be automatically created in the next version of PyScaffold (v4.x).
#
# Please have look at the docs for better alternatives
# (`Dependency Management` section).
# =============================================================================
#
# Add your pinned requirements so that they can be easily installed with:
# pip install -r requirements.txt
# Remember to also add them in setup.cfg but unpinned.
# Example:
# numpy==1.13.3
# scipy==1.0
#

#setuptool
setuptools==47.3.1
sphinx==3.1.1
wheel==0.34.2

#tests
pytest==5.4.3
pytest-cov==2.10.0
mock==4.0.2

#addition dependencyies
python-dotenv==0.14.0
discord.py[voice]==1.3.4